@@ -8,9 +8,9 @@
 			"type": "package",
 			"package": {
 				"name": "fortawesome/font-awesome",
-				"version": "6.1.1",
+				"version": "6.2.1",
 				"dist": {
-					"url": "https://github.com/FortAwesome/Font-Awesome/releases/download/6.1.1/fontawesome-free-6.1.1-web.zip",
+					"url": "https://github.com/FortAwesome/Font-Awesome/releases/download/6.2.1/fontawesome-free-6.2.1-web.zip",
 					"type": "zip"
 				}
 			}
@@ -19,9 +19,9 @@
 			"type": "package",
 			"package": {
 				"name": "angel-vladov/select2-theme-bootstrap4",
-				"version": "1.0.1",
+				"version": "1.0.2",
 				"dist": {
-					"url": "https://github.com/angel-vladov/select2-theme-bootstrap4/archive/1.0.1.zip",
+					"url": "https://github.com/angel-vladov/select2-theme-bootstrap4/archive/1.0.2.zip",
 					"type": "zip"
 				}
 			}
@@ -53,31 +53,31 @@
 		"ext-posix": "*",
 		"ext-simplexml": "*",
 		"ext-zip": "*",
-		"angel-vladov/select2-theme-bootstrap4": "1.0.1",
+		"angel-vladov/select2-theme-bootstrap4": "1.0.2",
 		"components/flag-icon-css": "^3.5",
 		"components/jquery": "^3.6",
 		"datatables/datatables": "^1.10",
-		"doctrine/doctrine-bundle": "^2.6",
+		"doctrine/doctrine-bundle": "^2.8",
 		"doctrine/doctrine-fixtures-bundle": "^3.4",
 		"doctrine/doctrine-migrations-bundle": "^3.2",
-		"doctrine/orm": "^2.12",
+		"doctrine/orm": "^2.14",
 		"eligrey/filesaver": "2.*",
 		"endclothing/prometheus_client_php": "^1.0",
 		"fortawesome/font-awesome": "6.*",
-		"friendsofsymfony/rest-bundle": "^3.3",
+		"friendsofsymfony/rest-bundle": "^3.5",
 		"ircmaxell/password-compat": "*",
-		"jms/serializer-bundle": "^4.0",
+		"jms/serializer-bundle": "^5.2",
 		"league/commonmark": "^2.3",
 		"mbostock/d3": "^3.5",
 		"myits/openid-connect-client": "*",
-		"nelmio/api-doc-bundle": "^4.9",
+		"nelmio/api-doc-bundle": "^4.11",
 		"novus/nvd3": "^1.8",
-		"phpunit/phpunit": "~9.5",
+		"phpunit/phpunit": "^9.6",
 		"ramsey/uuid": "^4.2",
 		"sebastian/diff": "^4.0",
 		"select2/select2": "4.*",
 		"sensio/framework-extra-bundle": "^6.2",
-		"sentry/sentry-symfony": "^4.3",
+		"sentry/sentry-symfony": "^4.5",
 		"symfony/asset": "5.4.*",
 		"symfony/browser-kit": "5.4.*",
 		"symfony/console": "5.4.*",
@@ -90,7 +90,7 @@
 		"symfony/http-client": "5.4.*",
 		"symfony/intl": "5.4.*",
 		"symfony/mime": "5.4.*",
-		"symfony/monolog-bundle": "^3.4.0",
+		"symfony/monolog-bundle": "^3.8.0",
 		"symfony/security-bundle": "5.4.*",
 		"symfony/security-csrf": "5.4.*",
 		"symfony/stopwatch": "5.4.*",
@@ -99,10 +99,10 @@
 		"symfony/web-profiler-bundle": "5.4.*",
 		"symfony/yaml": "5.4.*",
 		"twbs/bootstrap": "^4.6.0",
-		"twig/extra-bundle": "^3.4",
-		"twig/markdown-extra": "^3.4",
-		"twig/string-extra": "^3.4",
-		"twig/twig": "^3.4"
+		"twig/extra-bundle": "^3.5",
+		"twig/markdown-extra": "^3.5",
+		"twig/string-extra": "^3.5",
+		"twig/twig": "^3.5"
 	},
 	"require-dev": {
 		"ext-dom": "*",
