@@ -18,6 +18,11 @@ $domjudge_config = [];
 
 function judging_directory(string $workdirpath, array $judgeTask) : string
 {
+    if (filter_var($judgeTask['submitid'], FILTER_VALIDATE_INT) === false ||
+        filter_var($judgeTask['jobid'], FILTER_VALIDATE_INT) === false) {
+        error("Malformed data returned in judgeTask IDs");
+    }
+
     return $workdirpath . '/'
         . $judgeTask['submitid'] . '/'
         . $judgeTask['jobid'];
@@ -527,6 +532,13 @@ if (isset($options['verbose'])) {
 $runuser = RUNUSER;
 if (isset($options['daemonid'])) {
     $runuser .= '-' . $options['daemonid'];
+}
+
+if ($runuser === posix_getpwuid(posix_geteuid())['name'] ||
+    RUNGROUP === posix_getgrgid(posix_getegid())['name']
+) {
+    echo "Do not run the judgedaemon as the runser or rungroup.\n";
+    exit(1);
 }
 
 // Set static environment variables for passing path configuration
@@ -1264,11 +1276,11 @@ function judge(array $judgeTask): bool
         error("Could not create directory '$programdir'");
     }
 
-    // Cannot escape as we glob.
-    $cp_cmd = "cp -PRl '$workdir'/compile/* '$programdir'";
-    system($cp_cmd, $retval);
-    if ($retval!==0) {
-        error("Could not copy program to '$programdir'");
+    foreach (glob("$workdir/compile/*") as $compile_file) {
+        system('cp -PRl ' . dj_escapeshellarg($compile_file) . ' ' . dj_escapeshellarg($programdir), $retval);
+        if ($retval!==0) {
+            error("Could not copy program to '$programdir'");
+        }
     }
 
     // do the actual test-run
@@ -1339,9 +1351,15 @@ function judge(array $judgeTask): bool
     }
 
     if ($result === 'compare-error') {
-        logmsg(LOG_ERR, "comparing failed for compare script '" . $judgeTask['compare_script_id'] . "'");
-        $description = 'compare script ' . $judgeTask['compare_script_id'] . ' crashed';
-        disable('compare_script', 'compare_script_id', $judgeTask['compare_script_id'], $description, $judgeTask['judgetaskid']);
+        if ($combined_run_compare) {
+            logmsg(LOG_ERR, "comparing failed for combined run/compare script '" . $judgeTask['run_script_id'] . "'");
+            $description = 'combined run/compare script ' . $judgeTask['run_script_id'] . ' crashed';
+            disable('run_script', 'run_script_id', $judgeTask['run_script_id'], $description, $judgeTask['judgetaskid']);
+        } else {
+            logmsg(LOG_ERR, "comparing failed for compare script '" . $judgeTask['compare_script_id'] . "'");
+            $description = 'compare script ' . $judgeTask['compare_script_id'] . ' crashed';
+            disable('compare_script', 'compare_script_id', $judgeTask['compare_script_id'], $description, $judgeTask['judgetaskid']);
+        }
         return false;
     }
 
