@@ -9,12 +9,14 @@ import json
 import os
 import requests
 import requests.utils
+import subprocess
 import sys
 
 _myself = os.path.basename(sys.argv[0])
 _default_user_agent = requests.utils.default_user_agent()
 headers = {'user-agent': f'dj_utils/{_myself} ({_default_user_agent})'}
-api_url = 'unset'
+domjudge_webapp_folder_or_api_url = 'unset'
+ca_check = True
 
 
 def confirm(message: str, default: bool) -> bool:
@@ -38,12 +40,18 @@ def parse_api_response(name: str, response: requests.Response):
             raise RuntimeError(
                 f'API request {name} failed (code {response.status_code}).')
 
+    if response.status_code == 204:
+        return None
+
     # We got a successful HTTP response. It worked. Return the full response
     return json.loads(response.text)
 
 
 def do_api_request(name: str):
     '''Perform an API call to the given endpoint and return its data.
+
+    Based on whether `domjudge_webapp_folder_or_api_url` is a folder or URL this
+    will use the DOMjudge CLI or HTTP API.
 
     Parameters:
         name (str): the endpoint to call
@@ -55,18 +63,32 @@ def do_api_request(name: str):
         RuntimeError when the response is not JSON or the HTTP status code is non 2xx.
     '''
 
-    url = f'{api_url}/{name}'
+    if os.path.isdir(domjudge_webapp_folder_or_api_url):
+        return api_via_cli(name)
+    else:
+        global ca_check
+        url = f'{domjudge_webapp_folder_or_api_url}/{name}'
 
-    try:
-        response = requests.get(url, headers=headers)
-    except requests.exceptions.RequestException as e:
-        raise RuntimeError(e)
-
+        try:
+            response = requests.get(url, headers=headers, verify=ca_check)
+        except requests.exceptions.SSLError as e:
+            ca_check = not confirm(
+                "Can not verify certificate, ignore certificate check?", False)
+            if ca_check:
+                print('Can not verify certificate chain for DOMserver.')
+                exit(1)
+            else:
+                return do_api_request(name)
+        except requests.exceptions.RequestException as e:
+            raise RuntimeError(e)
     return parse_api_response(name, response)
 
 
 def upload_file(name: str, apifilename: str, file: str, data: dict = {}):
     '''Upload the given file to the API at the given path with the given name.
+
+    Based on whether `domjudge_webapp_folder_or_api_url` is a folder or URL this
+    will use the DOMjudge CLI or HTTP API.
 
     Parameters:
         name (str): the endpoint to call
@@ -80,10 +102,66 @@ def upload_file(name: str, apifilename: str, file: str, data: dict = {}):
         RuntimeError when the HTTP status code is non 2xx.
     '''
 
-    files = [(apifilename, open(file, 'rb'))]
+    if os.path.isdir(domjudge_webapp_folder_or_api_url):
+        return api_via_cli(name, 'POST', data, {apifilename: file})
+    else:
+        global ca_check
+        files = [(apifilename, open(file, 'rb'))]
 
-    url = f'{api_url}/{name}'
+        url = f'{domjudge_webapp_folder_or_api_url}/{name}'
 
-    response = requests.post(url, files=files, headers=headers, data=data)
+        try:
+            response = requests.post(
+                url, files=files, headers=headers, data=data, verify=ca_check)
+        except requests.exceptions.SSLError as e:
+            ca_check = not confirm(
+                "Can not verify certificate, ignore certificate check?", False)
+            if ca_check:
+                print('Can not verify certificate chain for DOMserver.')
+                exit(1)
+            else:
+                response = requests.post(
+                    url, files=files, headers=headers, data=data, verify=ca_check)
 
     return parse_api_response(name, response)
+
+
+def api_via_cli(name: str, method: str = 'GET', data: dict = {}, files: dict = {}):
+    '''Perform the given API request using the CLI
+
+    Parameters:
+        name (str): the endpoint to call
+        method (str): the method to use. Either GET or POST
+        data (dict): the POST data to use. Only used when method is POST
+        files (dict): the files to use. Only used when method is POST
+
+    Returns:
+        The parsed endpoint contents.
+
+    Raises:
+        RuntimeError when the command exit code is not 0.
+    '''
+
+    command = [
+        f'{domjudge_webapp_folder_or_api_url}/bin/console',
+        'api:call',
+        '-m',
+        method
+    ]
+
+    for item in data:
+        command.extend(['-d', f'{item}={data[item]}'])
+
+    for item in files:
+        command.extend(['-f', f'{item}={files[item]}'])
+
+    command.append(name)
+
+    result = subprocess.run(command, capture_output=True)
+    response = result.stdout.decode('ascii')
+
+    if result.returncode != 0:
+        print(response)
+        raise RuntimeError(f'API request {name} failed')
+
+    return json.loads(response)
