@@ -152,7 +152,7 @@ abstract class AbstractRestController extends AbstractFOSRestController
             $params['cid'] = $request->attributes->get('cid');
         }
         $headers = [
-            'Location' => $this->generateUrl("v4_app_api_${routeType}_single", $params, UrlGeneratorInterface::ABSOLUTE_URL),
+            'Location' => $this->generateUrl("v4_app_api_{$routeType}_single", $params, UrlGeneratorInterface::ABSOLUTE_URL),
         ];
         return $this->renderData($request, $data, Response::HTTP_CREATED,
             $headers);
