@@ -31,6 +31,7 @@ use Symfony\Component\Routing\RouterInterface;
 
 /**
  * @OA\Tag(name="General")
+ * @OA\Parameter(ref="#/components/parameters/strict")
  * @OA\Response(response="400", ref="#/components/responses/InvalidResponse")
  * @OA\Response(response="401", ref="#/components/responses/Unauthenticated")
  * @OA\Response(response="403", ref="#/components/responses/Unauthorized")
@@ -113,19 +114,25 @@ class GeneralInfoController extends AbstractFOSRestController
      *     )
      * )
      */
-    public function getInfoAction(): array
+    public function getInfoAction(Request $request): array
     {
-        return [
+        $strict = $request->query->getBoolean('strict', false);
+
+        $result = [
             'version' => self::CCS_SPEC_API_VERSION,
             'version_url' => self::CCS_SPEC_API_URL,
             'name' => 'DOMjudge',
-            'domjudge' => [
+        ];
+        if (!$strict) {
+            $result['domjudge'] = [
                 'api_version' => static::API_VERSION,
                 'version' => $this->getParameter('domjudge.version'),
                 'environment' => $this->getParameter('kernel.environment'),
                 'doc_url' => $this->router->generate('app.swagger_ui', [], RouterInterface::ABSOLUTE_URL),
-            ],
-        ];
+            ];
+        }
+
+        return $result;
     }
 
     /**
@@ -304,6 +311,12 @@ class GeneralInfoController extends AbstractFOSRestController
      *     description="The ISO 3166-1 alpha-3 code for the country to get the flag for",
      *     @OA\Schema(type="string")
      * )
+     * @OA\Parameter(
+     *     name="size",
+     *     in="path",
+     *     description="Preferred aspect ratio as <int>x<int>, currently only 1x1 and 4x3 are available.",
+     *     @OA\Schema(type="string")
+     * )
      */
     public function countryFlagAction(Request $request, string $countryCode, string $size): Response
     {
@@ -319,6 +332,9 @@ class GeneralInfoController extends AbstractFOSRestController
         $alpha3code = strtoupper($countryCode);
         if (!Countries::alpha3CodeExists($alpha3code)) {
             throw new NotFoundHttpException("country $alpha3code does not exist");
+        }
+        if (!preg_match('/^\d+x\d+$/', $size)) {
+            throw new BadRequestHttpException('invalid format for size parameter, should be "4x3" or "1x1"');
         }
         $alpha2code = strtolower(Countries::getAlpha2Code($alpha3code));
         $flagFile = sprintf('%s/public/flags/%s/%s.svg', $this->dj->getDomjudgeWebappDir(), $size, $alpha2code);
