@@ -22,6 +22,7 @@ use Symfony\Component\HttpKernel\Exception\BadRequestHttpException;
  * @Rest\Route("/contests/{cid}/runs")
  * @OA\Tag(name="Runs")
  * @OA\Parameter(ref="#/components/parameters/cid")
+ * @OA\Parameter(ref="#/components/parameters/strict")
  * @OA\Response(response="400", ref="#/components/responses/InvalidResponse")
  * @OA\Response(response="401", ref="#/components/responses/Unauthenticated")
  * @OA\Response(response="403", ref="#/components/responses/Unauthorized")
@@ -65,7 +66,6 @@ class RunController extends AbstractRestController implements QueryObjectTransfo
      *     )
      * )
      * @OA\Parameter(ref="#/components/parameters/idlist")
-     * @OA\Parameter(ref="#/components/parameters/strict")
      * @OA\Parameter(
      *     name="first_id",
      *     in="query",
@@ -101,7 +101,7 @@ class RunController extends AbstractRestController implements QueryObjectTransfo
      * Get the given run for this contest.
      * @throws NonUniqueResultException
      * @Security("is_granted('ROLE_JURY') or is_granted('ROLE_JUDGEHOST') or is_granted('ROLE_API_READER')")
-     * @Rest\Get("/{id}")
+     * @Rest\Get("/{id<\d+>}")
      * @OA\Response(
      *     response="200",
      *     description="Returns the given run for this contest",
@@ -113,7 +113,6 @@ class RunController extends AbstractRestController implements QueryObjectTransfo
      *     )
      * )
      * @OA\Parameter(ref="#/components/parameters/id")
-     * @OA\Parameter(ref="#/components/parameters/strict")
      */
     public function singleAction(Request $request, string $id): Response
     {
