@@ -177,7 +177,7 @@ class TwigExtension extends AbstractExtension implements GlobalsInterface
 
     public function printremainingminutes(float $start, float $end): string
     {
-        $minutesRemaining = floor(($end - $start)/60, 0);
+        $minutesRemaining = floor(($end - $start)/60);
         if ($minutesRemaining < 1) {
             return 'less than 1 minute to go';
         } elseif ($minutesRemaining == 1) {
@@ -1143,7 +1143,7 @@ EOF;
                         '<td><code>%s</code></td>',
                         $diff['external'] === null ? $null : $diff['external']
                     );
-                    $rows[]     = "<tr>${tdField}${tdUs}${tdExternal}</tr>";
+                    $rows[]     = "<tr>{$tdField}{$tdUs}{$tdExternal}</tr>";
                 }
 
                 $header  = <<<'EOF'
