@@ -106,7 +106,8 @@ class GeneralInfoControllerTest extends BaseTest
         static::assertEquals($teamname, $response['team']);
         static::assertEquals($roles, $response['roles']);
         static::assertTrue($response['enabled']);
-        static::assertGreaterThanOrEqual($response['first_login_time'], $response['last_login_time']);
+        static::assertNull($response['last_login_time']);
+        static::assertGreaterThanOrEqual($response['first_login_time'], $response['last_api_login_time']);
         $keysExpected = ['id', 'ip', 'last_ip', 'email'];
         foreach ($keysExpected as $keyExpected) {
             static::assertArrayHasKey($keyExpected, $response);
